+++ app/src/main/java/com/example/unlibrary/exchange/ExchangeFragment.java
@@ -78,9 +78,10 @@ public class ExchangeFragment extends Fragment {
 
         // Setup observer
         mViewModel.getNavigationEvent().observe(this, navDirections -> Navigation.findNavController(mBinding.getRoot()).navigate(navDirections));
+
         mViewModel.getSuccessRequestMsgEvent().observe(this, s -> {
                     ((MainActivity) requireActivity()).showToast(s);
-                    mViewModel.generateNotification(this.getView());
+                    mViewModel.sendRequestNotification(this.getView());
         });
 
         return mBinding.getRoot();
